+++ src/main/java/pl/app/wardrobe/clothes/view/ClothesList.java
@@ -3,24 +3,29 @@ package pl.app.wardrobe.clothes.view;
 import jakarta.enterprise.context.RequestScoped;
 import jakarta.inject.Inject;
 import jakarta.inject.Named;
+import pl.app.wardrobe.clothes.entity.Item;
 import pl.app.wardrobe.clothes.model.ClothesListModel;
 import pl.app.wardrobe.clothes.model.ItemListModel;
 import pl.app.wardrobe.clothes.service.ClothesService;
 import pl.app.wardrobe.clothes.service.ItemService;
+import pl.app.wardrobe.controller.servlet.exception.NotFoundException;
 import pl.app.wardrobe.factory.ModelFunctionFactory;
+
+import java.util.List;
 
 @RequestScoped
 @Named
 public class ClothesList {
     private final ClothesService clothesService;
-
+    private final ItemService itemService;
     private ClothesListModel clothes_list;
 
     private final ModelFunctionFactory factory;
 
     @Inject
-    public ClothesList(ClothesService clothesService, ModelFunctionFactory factory) {
+    public ClothesList(ClothesService clothesService,  ItemService itemService, ModelFunctionFactory factory) {
         this.clothesService = clothesService;
+        this.itemService = itemService;
         this.factory = factory;
     }
 
@@ -33,7 +38,21 @@ public class ClothesList {
     }
 
     public String deleteAction(ClothesListModel.Clothes clothes) {
-        clothesService.delete(clothes.getId());
+        clothesService.findClothesById(clothes.getId()).ifPresentOrElse(
+                entity -> {
+                    List<Item> items = itemService.findItemsByClothes(clothes.getId());
+                    if (items.isEmpty()) {
+                        throw new NotFoundException();
+                    }
+                    items.forEach(item -> itemService.delete(item.getId()));
+
+                    clothesService.delete(clothes.getId());
+                },
+                () -> {
+                    throw new NotFoundException();
+                }
+        );
+
         return "clothes_list?faces-redirect=true";
     }
 }
